Fix compiler for release 67+
Remove glibc223 package

Add more libraries to symlink

Check to make sure the files exists before symlinking

Build armv7l without selinux

Add pre-built binaries
